chore(ui-admin): rename modules to reducers
