+++ src/bp/ui-admin/src/reducers/index.js
