+++ src/bp/ui-admin/src/reducers/license.js
