+++ src/bp/ui-admin/src/reducers/versioning.js
